% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_xyz.R
\name{array_to_xyz}
\alias{array_to_xyz}
\title{Convert an Array to xyz Coordinates}
\usage{
array_to_xyz(array)
}
\arguments{
\item{array}{an 2 or 3-dimensional array}
}
\value{
matrix
}
\description{
This function takes a 2 or 3 dimensional image array and returns a matrix
containing values in the array along with their xyz coordinates within the
array. The xyz coordinates of a value will correspond to its indices in the
original array. This is useful for assigning xyz coordinates to values in a
raster image in order to find spacial data.
}
\examples{
test_array <- array(c(1,2,3,4), dim = c(2,2))
array_to_xyz(test_array)
}
