% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_array.R
\name{crop_array}
\alias{crop_array}
\title{Crops an Image Array}
\usage{
crop_array(image_array, value = 0)
}
\arguments{
\item{image_array}{a 3-dimensional image array}

\item{value}{maximum value that cropped matrices can contain}
}
\value{
array
}
\description{
This function crops an image array by removing all matrices on outer edges
that do not contain any entries greater than a specified value.
}
