% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_prop_comp.R
\name{hist_prop_comp}
\alias{hist_prop_comp}
\title{Proportion of Overlap Between Image Grayscale Distributions}
\usage{
hist_prop_comp(x, y, to = 2^16, by = 2^4)
}
\arguments{
\item{x}{An image array}

\item{y}{An image array}

\item{to}{Maximum value included in grayscale distribution}

\item{by}{Size of bins in the grayscale distribution}
}
\value{
A Data Frame
}
\description{
This function compares the grayscale distributions of two images. It finds
the histogram for the two images at a given bin size and then compares the
proportion of difference between the two.
}
