% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_cluster.R
\name{image_cluster}
\alias{image_cluster}
\title{Finds Connected Clusters Within an Image}
\usage{
image_cluster(
  x,
  max_dist = 10,
  min_intensity = 2000,
  min_cluster_size = 50,
  sample_prop = 2^(-8),
  sample_type = "grid"
)
}
\arguments{
\item{x}{An image array}

\item{max_dist}{Maximum distance between connected points}

\item{min_intensity}{Minimum intensity for points}

\item{min_cluster_size}{Minimum size of cluster in returned list}

\item{sample_prop}{Proportion of}

\item{sample_type}{Takes values "grid" or "random"}
}
\value{
List of clusters
}
\description{
This function finds clusters of points in an image and returns a list of points
in each cluster. Image arrays are first filtered, removing all points below the
set 'min_intensity'. Then, a connectivity graph is made with the remaining points
in the array after being given xyz coordinate corresponding to a points location
within the image array (this is done using the 'array_to_xyz() function). The
maximum distance between connected points in the graph is decided by the variable
'max_dist'. A list is then returned containing the points in each graph.
}
