% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_plot.R
\name{max_plot}
\alias{max_plot}
\title{Plot 3-Dimensional Images as 2-dimensional Projections}
\usage{
max_plot(array)
}
\arguments{
\item{array}{a 3-dimensional image array}
}
\value{
cimg plot
}
\description{
Plots a 3-dimensional grayscale raster image as three separate 2-dimensional
cimg plots. Each of these plots is the projection of maximum values from a
'face' of the original image onto a plane. These three images give us an
approximation of the original image that can be viewed in R.
}
